refactor(preprocessor): extract shared LLM logic from summarize and query
- Create src/core/llm/index.ts with shared LLM execution logic
- Create src/cli/utils.ts with shared loadScript function
- Remove duplicate code from summarize and query modules
- Delete provider.ts (merged into llm/index.ts)
